+++ cantera18/branches/scons/Cantera/src/thermo/MolalityVPSSTP.cpp
@@ -170,7 +170,7 @@ namespace Cantera {
   /*
    * return the solvent id index number.
    */
-  int MolalityVPSSTP::solventIndex() const {
+  size_t MolalityVPSSTP::solventIndex() const {
     return m_indexSolvent;
   }
